Move ast/grammar tests into the grammar package
